Hotfix - Corrige migracion montoMin y MontoMax en Actividad

+++ database/migrations/2018_08_13_121159_incluye_pago_minimo_y_maximo_en_actividad.php
@@ -28,8 +28,10 @@ class IncluyePagoMinimoYMaximoEnActividad extends Migration
      */
     public function down()
     {
-        Schema::table('Actividad', function (Blueprint $table) {
-            $table->dropColumn(['montoMin', 'montoMax']);
-        });
+        if (Schema::hasColumn('Persona', 'montoMin')) {
+            Schema::table('Actividad', function (Blueprint $table) {
+                $table->dropColumn(['montoMin', 'montoMax']);
+            });
+        }
     }
 }
